finished basic UI for profile page
