+++ client/src/pages/ProfilePage/ProfileView/Portfolio.jsx
@@ -6,10 +6,13 @@ import {
   forwardRef,
 } from "react";
 import { Button, Modal } from "flowbite-react";
-import { addPortfolio } from "./../../api/user";
+import { addPortfolio } from "./../../../api/user";
 import { useDispatch } from "react-redux";
-import { updateUser } from "./../../state";
+import { updateUser } from "./../../../state";
 import { useSelector } from "react-redux";
+import { MdModeEditOutline } from "react-icons/md";
+import { FaLinkedin, FaGithub } from "react-icons/fa";
+import { VscGlobe } from "react-icons/vsc";
 
 function Portfolio() {
   const user = useSelector((state) => state.auth.user);
@@ -57,8 +60,17 @@ function Portfolio() {
 
   return (
     <div className="widget_container">
-      <h2 className="font-bold text-lg">Portfolio</h2>
-      <Button onClick={() => setOpenModal(true)}>+</Button>
+      <div className="flex items-center justify-between text-primaryDark">
+        <div>
+          <h2 className="font-bold text-lg">Links</h2>
+        </div>
+        <button
+          onClick={() => setOpenModal(true)}
+          className="hover:text-primary cursor-pointer"
+        >
+          <MdModeEditOutline />
+        </button>
+      </div>
       <Modal show={openModal} onClose={() => setOpenModal(false)}>
         <Modal.Header>Add Portfolio</Modal.Header>
         <Modal.Body>
@@ -129,9 +141,39 @@ function Portfolio() {
         <>
           {portfolio.map((item, index) => (
             <div key={index}>
-              <h3>{item.linkedin}</h3>
-              <h3>{item.github}</h3>
-              <h3>{item.website}</h3>
+              {item.linkedin && (
+                <a
+                  href={`https://${item.linkedin}`}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="flex items-center gap-1"
+                >
+                  <FaLinkedin />
+                  {item.linkedin}
+                </a>
+              )}
+              {item.github && (
+                <a
+                  href={`https://${item.github}`}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="flex items-center gap-1"
+                >
+                  <FaGithub />
+                  {item.github}
+                </a>
+              )}
+              {item.website && (
+                <a
+                  href={`https://${item.website}`}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="flex items-center gap-1"
+                >
+                  <VscGlobe />
+                  {item.website}
+                </a>
+              )}
             </div>
           ))}
         </>
